+++ src/App.js
@@ -1,10 +1,9 @@
 import "./App.css";
-import Proposal6 from "./pages/Proposal6";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import LoadingPage from "./pages/LoadingPage";
 import LoginHome from "./pages/LoginHome";
 import MainHome from "./pages/MainHome";
-import ProposalSize from "./pages/ProposalSize";
+import ProposalPrice from "./pages/Proposal/ProposalPrice";
 
 function App() {
   return (
@@ -12,7 +11,7 @@ function App() {
       <Routes>
         <Route exact path="/loginhome" element={<LoginHome />} />
         <Route exact path="/loading" element={<LoadingPage />} />
-        <Route exact path="/proposal" element={<Proposal6 />} />
+        <Route exact path="/proposal" element={<ProposalPrice />} />
         <Route exact path="/home" element={<MainHome />} />
       </Routes>
     </BrowserRouter>
