+++ src/pages/Proposal/Proposal6.js
